# add benchmarks to compare protocompile perf and memory usage against other compilers