--- a/go.work
+++ b/go.work
@@ -1,0 +1,7 @@
+go 1.19
+
+use (
+	.
+	./internal/benchmarks
+	./internal/tools
+)