--- a/internal/benchmarks/go.mod
+++ b/internal/benchmarks/go.mod
@@ -1,0 +1,21 @@
+module github.com/jhump/protocompile/internal/benchmarks
+
+go 1.19
+
+require (
+	github.com/bufbuild/protocompile v0.0.0-20221004230924-06a336f5b6be
+	github.com/igrmk/treemap/v2 v2.0.1
+	github.com/jhump/protoreflect v1.13.0
+	github.com/stretchr/testify v1.8.0
+	google.golang.org/protobuf v1.28.2-0.20220831092852-f930b1dc76e8
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/golang/protobuf v1.5.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	golang.org/x/exp v0.0.0-20220317015231-48e79f11773a // indirect
+	golang.org/x/sync v0.0.0-20210220032951-036812b2e83c // indirect
+	google.golang.org/genproto v0.0.0-20200526211855-cb27e3aa2013 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)